test/py: Rework test_efi_bootmgr to not use virt-make-fs
The problem with using "virt-make-fs" to make a filesystem image is that
it is extremely slow. Switch to using the fs_helper functions we have
instead from the filesystem tests as these can add files to images and
are significantly faster and still do not require root access. In order
to complete this migration we also have to use "sfdisk" and "dd" to make
a full disk image rather than a single partition that we can use as-is.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
The default git settings don't detect the rename so well here because of
all of the spacing changes as well. I found I had to use "git show -M05
-w" in order to get something that confirmed the test itself wasn't
changed.

Cc: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Cc: Heinrich Schuchardt <xypron.glpk@gmx.de>