am33xx: am3517 evm: Drop old musb AM35X driver
The USB_AM35X is a legacy driver with one remaining user. The rest
of the am33xx devices use the USB_MUSB_TI driver for DM_USB so
move the last user USB_MUSB_TI which is the newer musb driver
So move to the newer driver so we can drop the old driver.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>